# specs/001-country-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Country WebAPI Service
  description: |
    High-performance, read-optimized Country WebAPI Service serving as the canonical source for country data
    across the Maliev microservices ecosystem. Features multi-layer caching, ETag-based conditional requests,
    and comprehensive administrative endpoints with optimistic concurrency control.
  version: 1.0.0
  contact:
    name: Maliev API Support
    email: api-support@maliev.com
servers:
  - url: https://api.maliev.com/country/v1
    description: Production
  - url: https://api-staging.maliev.com/country/v1
    description: Staging
  - url: https://api-dev.maliev.com/country/v1
    description: Development
  - url: http://localhost:5000/country/v1
    description: Local Development
...
paths:
...
  /admin/bulk-import:
    post:
      tags: [Bulk Import]
      summary: Submit bulk import job
      description: |
        Submits a bulk import job for validation. All records are validated atomically before
        any data is persisted. Returns job ID for status polling.
      operationId: submitBulkImport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkImportRequest'
      responses:
        '202':
          description: Import job accepted for processing
          headers:
            Location:
              schema:
                type: string
              example: /country/v1/admin/bulk-import/123
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportStatusResponse'
        '400':
          description: Validation failed - contains detailed errors per row
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportValidationErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: Payload too large (max 1000 countries per batch)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /admin/bulk-import/{jobId}:
    get:
      tags: [Bulk Import]
      summary: Get bulk import job status
      description: Returns current status of a bulk import job, including validation errors if any.
      operationId: getBulkImportStatus
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 123
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /admin/bulk-import/{jobId}/process:
    post:
      tags: [Bulk Import]
      summary: Process validated bulk import job
      description: |
        Triggers processing of a validated import job. Only jobs in "Validated" status can be processed.
        Processing is asynchronous - poll status endpoint for completion.
      operationId: processBulkImport
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: Processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportStatusResponse'
        '400':
          description: Job not in valid state for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ===== HEALTH ENDPOINTS =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Maliev Auth Service
...
  schemas:
...
    CreateCountryRequest:
      type: object
      required: [iso2, iso3, name]
      properties:
        iso2:
          type: string
          pattern: '^[A-Z]{2}$'
        iso3:
          type: string
          pattern: '^[A-Z]{3}$'
        name:
          type: string
          minLength: 1
          maxLength: 100
        officialName:
          type: string
          maxLength: 200
        numericCode:
          type: string
          pattern: '^[0-9]{3}$'
        capital:
          type: string
          maxLength: 100
        region:
          type: string
          maxLength: 50
        subregion:
          type: string
          maxLength: 50
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        demonym:
          type: string
          maxLength: 50
        areaKm2:
          type: number
          minimum: 0
        population:
          type: integer
          format: int64
          minimum: 0
        giniCoefficient:
          type: number
          minimum: 0
          maximum: 100
        timezones:
          type: array
          items:
            type: string
        borders:
          type: array
          items:
            type: string
            pattern: '^[A-Z]{3}$'
        callingCodes:
          type: array
          items:
            type: string
        topLevelDomains:
          type: array
          items:
            type: string
        currencies:
          type: object
        languages:
          type: object
        translations:
          type: object
        flags:
          type: object
        coatOfArms:
          type: object
        independent:
          type: boolean
        unMember:
          type: boolean
        landlocked:
          type: boolean
...
    BulkImportRequest:
      type: object
      required: [countries]
      properties:
        countries:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/CreateCountryRequest'
    BulkImportStatusResponse:
      type: object
      required: [jobId, status, totalRecords, processedRecords, createdAtUtc]
      properties:
        jobId:
          type: integer
          format: int64
          example: 123
        status:
          type: string
          enum: [Pending, Validating, Validated, ValidationFailed, Processing, Completed, Failed]
          example: Completed
        totalRecords:
          type: integer
          example: 250
        processedRecords:
          type: integer
          example: 250
        failedRecords:
          type: integer
          example: 0
        errorMessage:
          type: string
          nullable: true
        createdAtUtc:
          type: string
          format: date-time
        startedAtUtc:
          type: string
          format: date-time
          nullable: true
        completedAtUtc:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          format: int64
          nullable: true
          example: 1543
    BulkImportValidationErrorResponse:
      type: object
      required: [jobId, status, totalRecords, errors]
      properties:
        jobId:
          type: integer
          format: int64
        status:
          type: string
          enum: [ValidationFailed]
        totalRecords:
          type: integer
        errors:
          type: array
          items:
            type: object
            required: [rowNumber, field, message]
            properties:
              rowNumber:
                type: integer
                minimum: 1
              field:
                type: string
              message:
                type: string
          example:
            - rowNumber: 5
              field: iso2
              message: "Duplicate ISO2 code 'US' in batch"
            - rowNumber: 47
              field: name
              message: "Country name cannot exceed 100 characters"
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: ValidationFailed
        message:
          type: string
          example: One or more validation errors occurred
        details:
          type: object
          nullable: true
        correlationId:
          type: string
          format: uuid
          nullable: true
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found - resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Too Many Requests - rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...